Fix optimizer issue in training code
This pull request fixes an issue with the optimizer in the training code. The previous code was using a deprecated argument for decay in the Keras optimizer which was causing a ValueError. The issue has been fixed by removing the decay argument and using the default learning rate. The code has been tested and verified to be working as expected.